Replace Proposals nav bar toggle with feature toggle
